backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
celery==5.3.4
Django==5.2.8
django-celery-beat==2.5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
Flask==3.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.31.0
openai==1.12.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
